Began work
Began to fill in generation code for the empty cave and the caverns
dungeon.

Also i wrote a simple utility class to read the property files. Still i
need to add the main config file and the "save system" to access all the
files. What i have to lookup is how to create randomized caves (not
spherical ones, i want the shape to be blob like ..., also the caves er
not placed in a grid as the old mod does, i want it to place the caves
in random distances and angles)