Files
I began to work on the save format. As i mentioned earlier, i want to
use something similar to mojangs -nbt structure format so i began to
code it. Problem is: I don't know how to look up a blocks id (e.g.
minecraft:stone) and get its metadata(e.g. a stairs orientaion). I need
this to finish the saving method for the files (!)

If anyone knows how this can be accomplished, it would be  a great help
for me :)