Structure files!
I began with writing the core of the structure file system. atm it is
NOT functional. However, theoretically it can load and save dungeons,
but before that works i need someone to implement the loot chest system
and provide me a method to check wether a block is a loot chest or not!

Also i continued writing the generation code for the cavern dungeon. it
creates the caverns and randomizes their orientation. however, still not
functional as the system for loading the dungeon configurations is still
missing...

Added the Vector Utility classes from one of my other projects, it can
be used to rotate vectors around the x y and z axis