caverns
Added more functionality to the cavern dungeon. it theoretically now creates the cavern (code for generating them still missing) and connects them via tunnels.

Also it randomly chooses if it generates chests and what loot it is. WARNING: loot is the VANILLA Mineshaft, Dungeon, Stronghold and Blachsmith loot for normal chests and the endcity chest for the boss chest

Still, this doesn't have any functionality for now as of the missing system for registering dungeons out of their prop files....