Fixes & Dungeon Placers
Began to implement the dungeon placers. They kinda work but are still a
WIP prototype as the textures and the cooldown (so that it will only
place ONE dungeon are missing).
Also I fixed a big game crashing bug that had to do with the ELootTable
class not being able to instantiate properly, so please do not remove
the call of it in the "PostInit" section :)