Wall Parts
Wall Part for normal wall is finished, however, it should be put in a
separate thread so it does not crash the game. However, im not that
familar with forge. So i dont know how i can implement that cause i dont
know if forge allows modifying the world from a separate thread...