Going back to use the old generator
I could not fix the slowness of the new generator cause there is not
really a reason.

Somehow, this is faster:
for(BlockPos bp : blocks) {
  world.setBlockState(bp, BLOCK)
}

than iterating through a list that contains the keys of a map that holds
the blocks.

However, the key list is way smaller than the complete pos list in the
first case