Basically rewrote the entire dataIO package ObjectToCQONUtil has been renamed to CQONConversionUtil and completely rewritten to be more concise and to support arrays, collections, maps, etc. IO is now a subpackage of data, so as to provide a more suitable location for data manipulation utils that do not directly involve IO. PrimitiveManipulationUtil now exists and (for now) provides only a single method that enables one to easily determine if a given Object represents a primitive wrapper (created for use in CQONConversionUtil but separated so that it may be used elsewhere)
ProtectionHandler and ProtectedRegion have also been updated to utilize the new CQONConversionUtil, though this is mostly for testing purposes and still has some minor issues
Deserialization from CQON is still yet to be implemented, and thus the protection system will still not protect any regions generated prior to the most resent restart of the server/reloading of the world. This functionality is planned, and should be much more straightforward to implement due to the many improvements present in this commit.